Fixed batch_size arg type

+++ models/generate.py
@@ -98,7 +98,7 @@ def generate():
                                      help="Sample from top_k docs (Marginalized, ELBO)")
     decoding_group.add_argument('--docs_sampling_temperature', type=float, default=1,
                                      help="Temperature used for sampling docs (Marginalized, ELBO)")
-    decoding_group.add_argument('--batch_size', type=float, default=1,
+    decoding_group.add_argument('--batch_size', type=int, default=1,
                                 help="Number of source strings used at a time")
 
     Experiment.add_argument_group(parser)
